feat: inputHandler 구현
관심사 분리
